wip: set node indexes
Function to set index struct member of nodes, the heart of sort algorithm.
Index is the position of that node must be in stack a, to stack a be correctly
sorted, it is sequential and imutable number to your respective node.

The index start in 1 and end in get_list_size(). Node with lowest number
will have index 1, node with second smallest number will have index 2, ...,
node with largest number will have get_list_size().

If strigns in argv are sorted and whe iterate over it, the position in
argv will be the index to be assigned. All we have to do is search linked
list a node with have the the number we are in argv iteration.

The print_stack function is showing a pretty visualization of the
indexes.